+++ src/api-legacy/v1/borehole/editinglist.py
@@ -93,6 +93,7 @@ class ListEditingBorehole(Action):
                     restriction_until_bho,
                     'YYYY-MM-DD'
                 ) as restriction_until,
+                national_interest,
                 location_x_bho as location_x,
                 location_y_bho as location_y,
                 location_x_lv03_bho as location_x_lv03,