+++ src/client/src/commons/form/borehole/segments/boreholeDetailSegment.js
@@ -6,10 +6,16 @@ import DomainTree from "../../domain/tree/domainTree";
 import TranslationText from "../../translationText";
 import { NumericFormat } from "react-number-format";
 import { Form, Segment } from "semantic-ui-react";
+import {
+  FormControl,
+  RadioGroup,
+  FormControlLabel,
+  Radio,
+} from "@mui/material";
 import { parseIfString } from "../../formUtils";
 
 const BoreholeDetailSegment = props => {
-  const { size, borehole, updateChange, updateNumber, debug, t } = props;
+  const { size, borehole, updateChange, updateNumber } = props;
   return (
     <Segment>
       <Form autoComplete="off" error size={size}>
@@ -169,51 +175,42 @@ const BoreholeDetailSegment = props => {
           <label>
             <TranslationText id="groundwater" />
           </label>
-          <Form.Group inline>
-            <Form.Radio
-              checked={borehole.data.extended.groundwater === true}
-              label={t("common:yes")}
-              onChange={(e, d) => {
-                updateChange("extended.groundwater", true, false);
-              }}
-            />
-            <Form.Radio
-              checked={borehole.data.extended.groundwater === false}
-              label={t("common:no")}
-              onChange={(e, d) => {
-                updateChange("extended.groundwater", false, false);
-              }}
-            />
-            <Form.Radio
-              checked={borehole.data.extended.groundwater === null}
-              label={t("common:np")}
-              onChange={(e, d) => {
-                updateChange("extended.groundwater", null, false);
-              }}
-            />
-            {debug === true ? (
-              <div>
-                <div
-                  style={{
-                    color: "red",
-                  }}>
-                  trans=yes
-                </div>
-                <div
-                  style={{
-                    color: "red",
-                  }}>
-                  trans=no
-                </div>
-                <div
-                  style={{
-                    color: "red",
-                  }}>
-                  trans=np
-                </div>
-              </div>
-            ) : null}
-          </Form.Group>
+          <FormControl class="radio-group">
+            <RadioGroup
+              row
+              value={
+                borehole.data.extended.groundwater === true
+                  ? "TRUE"
+                  : borehole.data.extended.groundwater === false
+                  ? "FALSE"
+                  : "NULL"
+              }
+              onChange={e => {
+                let value =
+                  e.target.value === "TRUE"
+                    ? true
+                    : e.target.value === "FALSE"
+                    ? false
+                    : null;
+                updateChange("extended.groundwater", value, false);
+              }}>
+              <FormControlLabel
+                value="TRUE"
+                control={<Radio />}
+                label={<TranslationText id={"yes"} />}
+              />
+              <FormControlLabel
+                value="FALSE"
+                control={<Radio />}
+                label={<TranslationText id={"no"} />}
+              />
+              <FormControlLabel
+                value="NULL"
+                control={<Radio />}
+                label={<TranslationText id={"np"} />}
+              />
+            </RadioGroup>
+          </FormControl>
         </Form.Field>
         <Form.Field required>
           <label>